+++ build.gradle
@@ -50,7 +50,7 @@ checkstyle {
 }
 
 application {
-    mainClassName = "duke.Launcher"
+    mainClassName = "duke.Main"
 }
 
 shadowJar {
