+++ src/main/java/duke/Main.java
@@ -1,31 +1,13 @@
 package duke;
 
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
 /**
  * Main driver for the Duke class.
  */
-public class Main extends Application {
-    private Duke duke = new Duke("data/tasks.txt");
+public class Main  {
 
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            stage.setTitle("Duke+");
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) {
+        Launcher.launch();
     }
+
+
 }
